+++ crates/oxc_parser/src/ts/types.rs
@@ -616,7 +616,7 @@ impl<'a> Parser<'a> {
     }
 
     pub fn is_at_named_tuple_element(&mut self) -> bool {
-        let offset = usize::from(self.at(Kind::Dot3));
+        let offset = u8::from(self.at(Kind::Dot3));
         let has_colon = self.nth_at(offset + 1, Kind::Colon);
         let has_question_colon =
             self.nth_at(offset + 1, Kind::Question) && self.nth_at(offset + 2, Kind::Colon);
